Reformat and cleanup code
