+++ src/main/java/com/dnastack/gcp/model/Checksum.java
@@ -9,21 +9,21 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Checksum {
 
-	public void setType(Type type) {
-		this.type = type;
-	}
+    public void setType(Type type) {
+        this.type = type;
+    }
 
-	public enum Type {
-		md5("md5"), multipart_md5("multipart-md5"), S3("S3"), sha256("sha256"), sha512("sha512");
+    public enum Type {
+        md5("md5"), multipart_md5("multipart-md5"), S3("S3"), sha256("sha256"), sha512("sha512");
 
-		private String val;
+        private String val;
 
-		private Type(String val) {
-			this.val = val;
-		}
-	}
+        Type(String val) {
+            this.val = val;
+        }
+    }
 
-	private String checksum;
-	private Type type;
-	
+    private String checksum;
+    private Type type;
+
 }
