+++ src/main/java/com/dnastack/gcp/model/DosUrl.java
@@ -1,18 +1,18 @@
 package com.dnastack.gcp.model;
 
-import java.util.Map;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.Map;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class URL {
-	
-	private String url;
-	private Map<String, String> system_metadata;
-	private Map<String, String> user_metadata;
-	
+public class DosUrl {
+
+    private String url;
+    private Map<String, String> system_metadata;
+    private Map<String, String> user_metadata;
+
 }
