+++ src/test/java/com/dnastack/gcp/GCPWrapperTest.java
@@ -7,32 +7,28 @@ import junit.framework.TestSuite;
 /**
  * Unit test for simple App.
  */
-public class GCPWrapperTest 
-    extends TestCase
-{
+public class GCPWrapperTest
+        extends TestCase {
     /**
      * Create the test case
      *
      * @param testName name of the test case
      */
-    public GCPWrapperTest( String testName )
-    {
-        super( testName );
+    public GCPWrapperTest(String testName) {
+        super(testName);
     }
 
     /**
      * @return the suite of tests being tested
      */
-    public static Test suite()
-    {
-        return new TestSuite( GCPWrapperTest.class );
+    public static Test suite() {
+        return new TestSuite(GCPWrapperTest.class);
     }
 
     /**
      * Rigourous Test :-)
      */
-    public void testApp()
-    {
-        assertTrue( true );
+    public void testApp() {
+        assertTrue(true);
     }
 }
